requirements.txt
altgraph==0.17.4
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==25.9.23
humanfriendly==10.0
loguru==0.7.3
mpmath==1.3.0
Nuitka==2.7.6
numpy==2.3.3
onnxruntime==1.19.0
opencv-python==4.11.0.86
ordered-set==4.1.0
packaging==25.0
pefile==2023.2.7
protobuf==6.32.1
psutil==7.0.0
pycryptodomex==3.21.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.9
PyQt6==6.9.1
PyQt6-Charts==6.9.0
PyQt6-Charts-Qt6==6.9.2
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pyreadline3==3.5.4
pywin32-ctypes==0.2.3
PyYAML==6.0.3
shiboken6==6.9.2
sympy==1.14.0
win32_setctime==1.2.0
zstandard==0.23.0
